requirements.txt
# Core dependencies
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
werkzeug==2.3.7

# OpenAI
openai>=1.10.0

# Vector Database
pinecone==6.0.1

# Document Processing
pyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.3
numpy==1.24.3
markdown==3.5.1
python-magic==0.4.27
tiktoken==0.5.1

# Utilities
tqdm==4.66.1
requests==2.31.0
nltk==3.8.1
